feat: RayTraced AO in raster

+++ source/editor/raster/AoPass.h
@@ -44,7 +44,7 @@ public:
         param.ssao_max_distance = ui_config.ssao_max_distance;
         param.ssao_sample_count = ui_config.ssao_sample_count;
         param.ssao_radius       = ui_config.ssao_radius;
-        param.ao_mode           = ui_config.ao_mode;
+        param.ao_mode           = static_cast<uint32>(ui_config.ao_mode);
         param.input_image       = input_image;
         param.normal_tex        = context.textures.normal.handle;
         param.position_tex      = context.textures.position.handle;
